// src/types/formularioDeAvaliacaoP8sMindfulness.d.ts
export interface FormularioDeAvaliacaoP8sMindfulnessI {
  /** Auto-incrementing primary key */
  id: number;

  createdAt?: string;

  fullName?: string;
  birthDate?: string;
  satisfactionLevel?: string;

  likedMostAndTakeaways?: string;
  likedLeastSuggestions?: string;
  personalChange?: string;
  selfDifference?: string;
  knowledgeMomentsImportance?: string;
  tellAFriend?: string;
  benefits?: string;
  authorizeUse?: string;
}